Move base handler to separate module
